Make the `function` field in Value a pointer
The Function struct has grown a lot - it contains five pointers now.
This would negate the benefit of using a union inside the Value typedef,
which is that regardless of the type of the Value the type-specific data
all shares the same space in memory. So I've shrunk the space taken by a
Function struct in the Value, from five pointers to one.